git_statistics/requirements.txt
PyGithub
bs4